Reorganize PHP files, update namespaces
* Moved block includes into `blocks` subfolder.
* Moved component includes into `components` subfolder.
* Some functions in content.php were really components, so moved them
  to a new file in the `components` subfolder and updated their
  namespace.
* There are currently only three functions that seem like "utilities"
  or "helpers" that aren't components, so moved them to a file in a
  `utilities` subfolder and updated their namespace.
* Definitions seem fairly foundational to our whole blocks module,
  rather than some kind of specialized utility, so left them in the
  root includes folder.
