+++ public_html/wp-content/mu-plugins/blocks/views/organizer.php
@@ -1,7 +1,9 @@
 <?php
 namespace WordCamp\Blocks\Organizers;
+
 use WP_Post;
-use function WordCamp\Blocks\Shared\Content\{ get_all_the_content, render_item_title, render_item_content };
+use function WordCamp\Blocks\Components\{ render_item_title, render_item_content };
+use function WordCamp\Blocks\Utilities\{ get_all_the_content };
 
 defined( 'WPINC' ) || die();
